Use the talk path instead of just the name
Pass the path to the talk directory - e.g.
`just generate ./src/test-driven-drupal` instead of just the talk
name/slug.

This doesn't change where the files are sourced from or output, but this
makes it easier to use `just generate` as I can use tab completion when
executing the command - making it quicker and less prone to typos.